Create numInd.function
This function will take genetic inputs of QTL Variance (the amount a trait varies in a population), dominance additivty (how dominant/recessive the model is, or the behavior of the heterozygote), and the increaser allele frequency (determines the popularity of the trait in a population).  It will then randomly generate a genotype and associated phenotype.